[engraving] replaced QList to std. Part 11

+++ src/engraving/libmscore/stafftype.h
@@ -422,7 +422,7 @@ public:
     bool isDrumStaff() const { return _group == StaffGroup::PERCUSSION; }
     // static functions for font config files
     static std::vector<QString> fontNames(bool bDuration);
-    static bool fontData(bool bDuration, int nIdx, QString* pFamily, QString* pDisplayName, qreal* pSize, qreal* pYOff);
+    static bool fontData(bool bDuration, size_t nIdx, QString* pFamily, QString* pDisplayName, qreal* pSize, qreal* pYOff);
 
     static void initStaffTypes();
     static const std::vector<StaffType>& presets() { return _presets; }
